backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
psycopg2-binary>=2.9.9
django-redis>=5.4.0
redis>=5.0.0
gunicorn>=22.0.0
python-dotenv>=1.0.1
whitenoise>=6.6.0
